share COTS utils function
